docs for classes, renaming folder common widget
