+++ lib/data_repositories/place_repository.dart
@@ -1,10 +1,9 @@
 import 'dart:convert';
-
 import 'package:dio/dio.dart';
 import 'package:places/data_other/exceptions.dart';
 import 'package:places/data_repositories/mocks.dart';
 import 'package:places/domain_models/place.dart';
-import 'package:places/ui_my_app/my_app.dart';
+import 'package:places/main.dart';
 
 late final Dio dio;
