+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
+import 'package:places/ui_commons/platform_detector.dart';
+import 'package:provider/provider.dart';
 import 'package:places/data_repositories/geo_repository.dart';
 import 'package:places/data_repositories/place_repository.dart';
 import 'package:places/data_repositories/search_repository.dart';
@@ -7,18 +9,13 @@ import 'package:places/domain_entities/geo_entity.dart';
 import 'package:places/domain_entities/place_entity.dart';
 import 'package:places/domain_entities/search_entity.dart';
 import 'package:places/ui_blocs/my_bloc_observer.dart';
-import 'package:provider/provider.dart';
-
-import 'package:places/ui_my_app/my_app.dart';
-
+import 'package:places/my_app_and_routes.dart';
 import 'package:places/ui_interactors/select_category_interactor.dart';
 import 'package:places/ui_interactors/place_interactor.dart';
 import 'package:places/ui_interactors/settings_interactor.dart';
 import 'package:places/ui_interactors/filter_interactor.dart';
 import 'package:places/ui_interactors/search_interactor.dart';
-
 import 'package:places/data_repositories/settings_repository.dart';
-
 import 'package:places/domain_entities/settings_entity.dart';
 
 late final SettingsInteractor settingsInteractor;
@@ -43,7 +40,6 @@ void main() {
   final placeRepository = PlaceRepository();
   placeEntity = PlaceEntity(placeRepository: placeRepository);
 
-  // !!!
   filterInteractor = FilterInteractor();
   placeInteractor = PlaceInteractor();
 
@@ -85,7 +81,15 @@ class MyAppProvider extends StatelessWidget {
           create: (context) => SelectCategoryInteractor(),
         ),
       ],
-      child: const MyApp(),
+      child: const MyAppAndRoutes(),
     );
   }
 }
+
+PlatformDetector platformDetector = PlatformDetector();
+
+const bool isDebugMockImagesInPlaceOfHttp = false;
+const bool isDebugMockDataInPlaceOfHttp = false;
+
+const bool isDebugUserStay = true;
+const bool isDebugUserMove = false;
