+++ lib/ui/screens/add_place_screen/dialog_add_photo.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'package:places/ui_my_app/ui_image_paths.dart';
-import 'package:places/ui_my_app/ui_strings.dart';
+import 'package:places/ui_commons/ui_image_paths.dart';
+import 'package:places/ui_commons/ui_strings.dart';
 
+/// Диалог "Добавить фото"
 class DialogAddPhoto extends StatelessWidget {
   const DialogAddPhoto({Key? key}) : super(key: key);
