+++ lib/ui/screens/add_place_screen/widget_category_header.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+/// Подзаголовок категории
 class CategoryHeaderWidget extends StatelessWidget {
   const CategoryHeaderWidget(
     this.categoryName, {
