+++ lib/ui/screens/main_1_places_list/widget_my_sliver_app_bar.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui_my_app/ui_strings.dart';
-import 'package:places/ui_widgets/headers.dart';
+import 'package:places/ui_commons/ui_strings.dart';
+import 'package:places/ui_common_widgets/headers.dart';
 
+/// Сливер-заголовок для главного экрана
 class WidgetMySliverAppBar extends SliverPersistentHeaderDelegate {
   WidgetMySliverAppBar({
     required this.expandedHeight,
