+++ lib/ui/screens/main_3_wished_and_seen/widget_tab_favorite.dart
@@ -1,17 +1,18 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:places/ui_blocs/places_cubit.dart';
-import 'package:places/ui_my_app/ui_strings.dart';
+import 'package:places/ui_commons/ui_strings.dart';
 import 'package:provider/provider.dart';
 
-import 'package:places/ui_my_app/enums.dart';
-import 'package:places/ui_my_app/my_scroll_physics.dart';
-import 'package:places/ui_widgets/place_card.dart';
+import 'package:places/ui_commons/enums.dart';
+import 'package:places/ui_commons/my_scroll_physics.dart';
+import 'package:places/ui_common_widgets/place_card.dart';
 import 'package:places/ui/screens/main_3_wished_and_seen/widget_empty_list.dart';
 import 'package:places/ui/screens/place_details_screen/screen_place_details.dart';
 
 import 'package:places/domain_models/place.dart';
 
+/// Экран "Избранные и посещенные", вкладка "Избранные"
 class WidgetTabFavorite extends StatefulWidget {
   const WidgetTabFavorite({
     required this.favoritePlaces,
