+++ lib/ui/screens/main_screen/main_screen.dart
@@ -5,6 +5,10 @@ import 'package:places/ui/screens/main_2_map/screen_main_2_map.dart';
 import 'package:places/ui/screens/main_3_wished_and_seen/screen_main3_favorite_and_visited_route.dart';
 import 'package:places/ui/screens/main_4_settings/screen_main_4_settings.dart';
 
+/// ГЛАВНЫЙ ЭКРАН.
+/// Состоит из 4х частей - переход к ним
+/// осуществляется через нажатие на кнопкам в BottomNavigationBar.
+/// 
 class AllMainScreens extends StatefulWidget {
   const AllMainScreens({Key? key}) : super(key: key);
   @override
