+++ lib/ui/screens/onboarding_screen/onboarding_page_1.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui_my_app/ui_image_paths.dart';
+import 'package:places/ui_commons/ui_image_paths.dart';
 
+/// Страница "Добро пожаловать 1"
 class OnboardingPage1 extends StatelessWidget {
   const OnboardingPage1({
     Key? key,
