+++ lib/ui/screens/place_details_screen/screen_place_details.dart
@@ -2,13 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:places/ui_interactors/place_interactor.dart';
 import 'package:provider/provider.dart';
 
-import 'package:places/ui_my_app/ui_strings.dart';
-import 'package:places/ui_my_app/my_scroll_physics.dart';
+import 'package:places/ui_commons/ui_strings.dart';
+import 'package:places/ui_commons/my_scroll_physics.dart';
 import 'package:places/ui/screens/place_details_screen/widget_back_button.dart';
-import 'package:places/ui_widgets/my_image_widget.dart';
+import 'package:places/ui_common_widgets/my_image_widget.dart';
 
 import 'package:places/domain_models/place.dart';
 
+/// Экран "Детальная страница места"
 class ScreenPlaceDetails extends StatefulWidget {
   const ScreenPlaceDetails({
     required this.placeId,
