--- lib/ui_blocs/places_states.dart
@@ -1,3 +1,0 @@
-class PlacesStates{
-  
-}
