+++ lib/ui_common_widgets/my_image_widget.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui_my_app/my_app.dart';
+import 'package:places/main.dart';
 
-/// Для отладки - для того, чтобы работал flutter for web подменяет
-/// url картинки моковой картинкой
+/// Виджет-обертка для Image
+/// (нужен для того, чтобы в режиме отладки работал flutter for web: 
+/// подменяет url картинки моковой картинкой)
 class MyImageWidget extends StatelessWidget {
   const MyImageWidget({
     required this.url,
