+++ lib/ui_common_widgets/place_card.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui_my_app/enums.dart';
-import 'package:places/ui_my_app/ui_strings.dart';
-import 'package:places/ui_widgets/my_image_widget.dart';
-import 'package:places/ui_widgets/place_card_icons.dart';
+import 'package:places/ui_commons/enums.dart';
+import 'package:places/ui_commons/ui_strings.dart';
+import 'package:places/ui_common_widgets/my_image_widget.dart';
+import 'package:places/ui_common_widgets/place_card_icons.dart';
 
 import 'package:places/domain_models/place.dart';
 
+/// Карточка места
 class PlaceCard extends StatelessWidget {
   const PlaceCard({
     required this.place,
