+++ lib/ui_commons/platform_detector.dart
