--- lib/ui_my_app/routes.dart
@@ -1,7 +1,0 @@
-const String ROUTE_HOME = '/';
-const String ROUTE_FILTER = '/filter';
-const String ROUTE_ADD = '/add';
-const String ROUTE_SELECT_CATEGORY = '/select_category';
-const String ROUTE_SEARCH = '/search';
-const String ROUTE_ONBOARDING = '/onboarding';
-const String ROUTE_SPLASH = '/splash';
